List all vaults if there are more than 10
By default `list_vaults` method returns only 10 vaults. If there are more vaults - those were not listed.

Response object contains `Marker` value if there are more than 10 vaults in the account. The `Marker` value should be provided as parameter to `list_vaults` method to retrieve list of further vaults. Such requests needs to be repeated until `Marker` is not present in the response.